create invoice with api

+++ WebApplication2/Models/InvoiceDetail.cs
@@ -22,7 +22,7 @@ namespace WebApplication2.Models
 
         public Int16 amount { get; set; }
 
-        public Int16 price { get; set; }
+        public double price { get; set; }
 
         [DatabaseGenerated(DatabaseGeneratedOption.Computed)]
         public Double subTotal
@@ -41,7 +41,7 @@ namespace WebApplication2.Models
 
         public Product Product { get; set; }
 
-        public InvoiceDetail(short invoice, short product, short amount, short price, double subTotal)
+        public InvoiceDetail(short invoice, short product, short amount, double price, double subTotal)
         {
             this.invoice = invoice;
             this.product = product;
